requirements.txt
feedparser
PyYAML
SQLAlchemy
psycopg2-binary
requests
openai
APScheduler
Flask
# ASGI server and middleware for async support
starlette
uvicorn[standard]
langchain
langchain_openai
langchain_community
lang